0asdfadf

+++ src/main/java/com/kosa/project/mapper/MemberMapper.java
@@ -5,7 +5,7 @@ import com.kosa.project.domain.MemberVO;
 
 public interface MemberMapper {
     public Integer insert(MemberVO vo);
-    public MemberVO find(Long idx);
-    public int delete(Long idx);
+    public MemberVO find(Integer idx);
+    public int delete(Integer idx);
     public int update(MemberVO vo);
 }
