+++ src/main/java/com/kosa/project/service/MemberService.java
@@ -4,7 +4,7 @@ import com.kosa.project.domain.MemberVO;
 
 public interface MemberService {
 	public Integer insert(MemberVO vo);
-	public MemberVO find(Long idx);
-	public boolean remove(Long idx);
+	public MemberVO find(Integer idx);
+	public boolean remove(Integer idx);
     public boolean modify(MemberVO vo);
 }
